erro de tipos corrigido + return false

+++ projeto/src/Camera/perspective.cpp
@@ -9,7 +9,7 @@
 
 bool Perspective::GenerateRay(const int x, const int y, Ray *r, const float *cam_jitter) {
     Vector result;
-    Vector xy1 = {x,y,1};
+    Vector xy1 = {static_cast<float>(x),static_cast<float>(y),1};
     Vector RUF[3] = { Vector(c2w[0][1],c2w[0][2],c2w[0][2]),
                       Vector(c2w[1][1],c2w[1][2],c2w[1][2]), 
                       Vector(c2w[2][1],c2w[2][2],c2w[2][2])};
